test: add conversation service model tests
